fix(inbox): resolve all 10 review findings
P1 fixes:
- Delete dead CreateProjectPopover.tsx (never imported)
- Delete dead CreateProjectInline function from AppSidebar
- Extract InboxSidebarBody into own file (AppSidebar 560→334 lines)

P2 fixes:
- Replace all hardcoded #FF6600/#E65C00 with bg-primary/text-primary tokens
- Wrap InboxPage in ErrorBoundary (matches all other pages)
- Replace bare animate-shimmer div with XerusLoader component
- Remove unused FolderPlus import
- Add console.error in catch blocks for dev observability

P3 fixes:
- Replace hand-drawn SVG with Lucide MessageSquare icon
- Rename ChatSidebarFallback → ChatSidebarLoading

All files now under 400 lines:
  AppSidebar.tsx: 334, InboxSidebarBody.tsx: 174, page.tsx: 293
